+++ ui/src/api/kubernetes/migrationResourceBundle/constants.ts
@@ -1,0 +1,24 @@
+export const RELATED_CRD_PLURALS = [
+  'arraycreds',
+  'arraycredsmappings',
+  'bmconfigs',
+  'clustermigrations',
+  'esximigrations',
+  'esxisshcreds',
+  'migrationplans',
+  'migrations',
+  'migrationtemplates',
+  'networkmappings',
+  'openstackcreds',
+  'pcdclusters',
+  'pcdhosts',
+  'rdmdisks',
+  'rollingmigrationplans',
+  'storagemappings',
+  'vjailbreaknodes',
+  'vmwareclusters',
+  'vmwarecreds',
+  'vmwarehosts',
+  'vmwaremachines',
+  'volumeimageprofiles'
+] as const
