+++ ui/src/api/kubernetes/migrationResourceBundle/index.ts
@@ -1,0 +1,1 @@
+export { fetchMigrationResourceBundle } from './buildMigrationResourceBundle'
